PP-4616: Move tests for uk.gov.pay.publicauth.util into the same package
The tests for classes in this package were mistakenly put into the
similarly-named uk.gov.pay.publicauth.utils

Move them and tighten up class access levels.
