+++ src/main/java/uk/gov/pay/publicauth/util/ApplicationStartupDependentResourceChecker.java
@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 import java.time.Duration;
 import java.util.function.Consumer;
 
-public class ApplicationStartupDependentResourceChecker {
+class ApplicationStartupDependentResourceChecker {
 
     private static final int PROGRESSIVE_SECONDS_TO_WAIT = 5;
     private static final Logger logger = LoggerFactory.getLogger(ApplicationStartupDependentResourceChecker.class);
@@ -14,12 +14,12 @@ public class ApplicationStartupDependentResourceChecker {
     private final ApplicationStartupDependentResource applicationStartupDependentResource;
     private final Consumer<Duration> waiter;
 
-    public ApplicationStartupDependentResourceChecker(ApplicationStartupDependentResource applicationStartupDependentResource, Consumer<Duration> waiter) {
+    ApplicationStartupDependentResourceChecker(ApplicationStartupDependentResource applicationStartupDependentResource, Consumer<Duration> waiter) {
         this.applicationStartupDependentResource = applicationStartupDependentResource;
         this.waiter = waiter;
     }
 
-    public void checkAndWaitForResource() {
+    void checkAndWaitForResource() {
         long timeToWait = 0;
         while(!applicationStartupDependentResource.isAvailable()) {
             timeToWait += PROGRESSIVE_SECONDS_TO_WAIT;
